backend/requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
gunicorn>=21.2.0
sqlalchemy>=2.0.25
psycopg2-binary>=2.9.11
pydantic>=2.5.3
pydantic-settings>=2.1.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.2
python-multipart==0.0.6
python-socketio==5.11.0
aiofiles==23.2.1
pydantic-core>=2.14.6
